Update README.md
This project is a user interface developed for a Stratocore company exercise. Created with HTML, CSS, and JavaScript, this project aims to demonstrate my skills in front-end development and UX/UI design. The interface allows you to manage user information, customize the homepage, and features a dark mode functionality for improved accessibility.